requirements.txt
# ============================================
# DEPENDÊNCIAS DO SISTEMA DE AUDIÊNCIAS
# ============================================

# Automação Web
selenium==4.15.2
webdriver-manager==4.0.1

# Manipulação de Dados
pandas==2.1.3
numpy==1.26.2

# APIs Google (Sheets, Calendar)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Requisições HTTP
requests==2.31.0

# Variáveis de Ambiente
python-dotenv==1.0.0

# Retry Logic
tenacity==8.2.3

# Utilitários
python-dateutil==2.8.2
pytz==2023.3.post1
